optimize(mtd): user synchronize_mtd only for list of af as param
If there is a `id_acquisition_frameworks` parameter ; list of AF (Acquisition Frameworks) ; when calling `get_datasets` or `get_acquisition_frameworks_list`, then the user MTD sync is made only for metadata associated to the provided list of AF: all the AF in the list and associated DS (DataSets) are then retrieved.

NOTE: such a parameter is actually provided when an AF is unrolled from the Metadata module at the user's click and the list of associated DS needs to be retrieved. This commit should therefore greatly reduce the execution time for many of these operations, especially as the user has a lot of metadata to retrieve (for all AF).